+++ src/IsoCodesFactory.php
@@ -7,9 +7,11 @@ use Sokil\IsoCodes\Database\Countries;
 use Sokil\IsoCodes\Database\Currencies;
 use Sokil\IsoCodes\Database\HistoricCountries;
 use Sokil\IsoCodes\Database\Languages;
+use Sokil\IsoCodes\Database\LanguagesInterface;
 use Sokil\IsoCodes\Database\LanguagesPartitioned;
 use Sokil\IsoCodes\Database\Scripts;
 use Sokil\IsoCodes\Database\Subdivisions;
+use Sokil\IsoCodes\Database\SubdivisionsInterface;
 use Sokil\IsoCodes\Database\SubdivisionsPartitioned;
 
 /**
@@ -18,29 +20,41 @@ use Sokil\IsoCodes\Database\SubdivisionsPartitioned;
 class IsoCodesFactory
 {
     /**
+     * Database splits into partition files.
+     *
+     * Fetching some entry will load only little part of database.
+     * Loaded entries not stored statically.
+     *
+     * This scenario may be useful when just few entries need
+     * to be loaded, for example on web request when one entry fetched.
+     *
+     * This may require a log of file read operations.
+     */
+    public const OPTIMISATION_MEMORY = 1;
+
+    /**
+     * Entire database loaded into memory from single JSON file once.
+     *
+     * All entries created and stored into RAM. Next read of save
+     * entry will just return it without io operations with files and building objects.
+     *
+     * This scenario may be useful for daemons to decrease file operations,
+     * or when most entries will be fetched from database.
+     *
+     * This may require a lot of RAM for storing all entries.
+     */
+    public const OPTIMISATION_TIME = 2;
+
+    /**
      * Path to directory with databases
      *
      * @var string
      */
     private $baseDirectory;
 
-    /**
-     * @var AbstractDatabase[]
-     */
-    private $databases = [];
-
     public function __construct(?string $baseDirectory = null)
     {
         $this->baseDirectory = $baseDirectory;
-    }
-
-    private function getDatabase(string $className): AbstractDatabase
-    {
-        if (empty($this->databases[$className])) {
-            $this->databases[$className] = new $className($this->baseDirectory);
-        }
-
-        return $this->databases[$className];
     }
 
     /**
@@ -48,25 +62,28 @@ class IsoCodesFactory
      */
     public function getCountries(): Countries
     {
-        return $this->getDatabase(Countries::class);
-    }
-
-    /**
-     * ISO 3166-2
-     */
-    public function getSubdivisions(): Subdivisions
-    {
-        return $this->getDatabase(Subdivisions::class);
+        return new Countries($this->baseDirectory);
     }
 
     /**
      * ISO 3166-2
      *
-     * Loaded from bunch of database files instead of one single file
+     * @param int $optimisation One of self::OPTIMISATION_* constants
      */
-    public function getSubdivisionsPartitioned(): SubdivisionsPartitioned
+    public function getSubdivisions(int $optimisation = self::OPTIMISATION_MEMORY): SubdivisionsInterface
     {
-        return $this->getDatabase(SubdivisionsPartitioned::class);
+        switch ($optimisation) {
+            case self::OPTIMISATION_MEMORY:
+                $database = new SubdivisionsPartitioned($this->baseDirectory);
+                break;
+            case self::OPTIMISATION_TIME:
+                $database = new Subdivisions($this->baseDirectory);
+                break;
+            default:
+                throw new \InvalidArgumentException('Invalid optimisation specified');
+        }
+
+        return $database;
     }
 
     /**
@@ -74,7 +91,7 @@ class IsoCodesFactory
      */
     public function getHistoricCountries(): HistoricCountries
     {
-        return $this->getDatabase(HistoricCountries::class);
+        return new HistoricCountries($this->baseDirectory);
     }
 
     /**
@@ -82,7 +99,7 @@ class IsoCodesFactory
      */
     public function getScripts(): Scripts
     {
-        return $this->getDatabase(Scripts::class);
+        return new Scripts($this->baseDirectory);
     }
 
     /**
@@ -90,22 +107,27 @@ class IsoCodesFactory
      */
     public function getCurrencies(): Currencies
     {
-        return $this->getDatabase(Currencies::class);
+        return new Currencies($this->baseDirectory);
     }
 
     /**
      * ISO 639-3
+     *
+     * @param int $optimisation One of self::OPTIMISATION_* constants
      */
-    public function getLanguages(): Languages
+    public function getLanguages(int $optimisation = self::OPTIMISATION_MEMORY): LanguagesInterface
     {
-        return $this->getDatabase(Languages::class);
-    }
+        switch ($optimisation) {
+            case self::OPTIMISATION_MEMORY:
+                $database = new LanguagesPartitioned($this->baseDirectory);
+                break;
+            case self::OPTIMISATION_TIME:
+                $database = new Languages($this->baseDirectory);
+                break;
+            default:
+                throw new \InvalidArgumentException('Invalid optimisation specified');
+        }
 
-    /**
-     * ISO 639-3
-     */
-    public function getLanguagesPartitioned(): LanguagesPartitioned
-    {
-        return $this->getDatabase(LanguagesPartitioned::class);
+        return $database;
     }
 }